@@ -1,8 +1,3 @@
 module github.com/kardianos/rdb
 
-go 1.13
-
-require (
-	golang.org/x/net v0.0.0-20200904194848-62affa334b73
-	vitess.io/vitess v0.7.0
-)
+go 1.14
